services/login_backend/requirements.txt
Flask==2.2.2
gunicorn==20.1.0
Flask_SQLAlchemy==3.0.3
psycopg2-binary==2.9.5
Flask-Login==0.6.2
google-cloud-pubsub==2.16.0
spotipy==2.23.0
flask-cors==3.0.10
Flask-Session==0.4.1